backend/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
numpy==2.2.1
pandas==2.2.3
scikit_learn==1.6.1
gunicorn
python-dotenv


